+++ invenio/modules/workflows/static/js/workflows/entry_details.js
@@ -17,13 +17,12 @@
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
  */
 
-var url_entry_data_preview = "/admin/bibworkflow/entry_data_preview";
 
-if ( window.addEventListener ) {
-        $("div.btn-group[name='object_preview_btn']").bind('click', function(event){
+function bind_object_preview(url_prefix) {
+    $("div.btn-group[name='object_preview_btn']").bind('click', function(event){
           var format = event.target.name;
           jQuery.ajax({
-                url: url_entry_data_preview + "?oid={{ entry.id }}&format=" + format,
+                url: url_prefix + "?oid={{ entry.id }}&format=" + format,
                 success: function(json){
                     if(format == 'xm' || format == 'marcxml'){
                         $('div[name="object_preview"]').wrapAll('<debug>').text(json);
