+++ invenio/modules/workflows/tasks/simplified_data_tasks.py
@@ -17,7 +17,7 @@
 
 """Basic simplified data test functions - NOT FOR XML """
 
-from invenio.bibworkflow_config import CFG_OBJECT_STATUS
+from ..config import CFG_OBJECT_STATUS
 
 
 def task_a(a):
