+++ invenio/modules/workflows/templates/workflows/layout.html
@@ -5,13 +5,13 @@
         <div class="span2">
             <div class="well sidebar-nav">
                 <ul class="nav nav-list">
-                    <li><a href="{{ url_for('bibworkflow.index') }}"
+                    <li><a href="{{ url_for('workflows.index') }}"
                         {% if request.path == url_for('workflows.index') %}
                             class="active"
                         {% endif %}>
                         Running workflows
                     </a></li>
-                    <li><a href="{{ url_for('bibworkflow.show_workflows') }}"
+                    <li><a href="{{ url_for('workflows.show_workflows') }}"
                     {% if request.path == url_for('workflows.show_workflows') %}
                         class="active"
                     {% endif %}>
