+++ invenio/modules/workflows/views/admin.py
@@ -25,9 +25,9 @@ from invenio.base.i18n import _
 from invenio.base.decorators import wash_arguments, templated
 from invenio.ext.breadcrumb import default_breadcrumb_root, register_breadcrumb
 from ..api import start_delayed
-from invenio.bibworkflow_utils import (get_workflow_definition,
-                                       get_redis_keys as utils_get_redis_keys,
-                                       filter_holdingpen_results)
+from ..utils import (get_workflow_definition,
+                     get_redis_keys as utils_get_redis_keys,
+                     filter_holdingpen_results)
 
 from ..models import Workflow, BibWorkflowObject
 from ..loader import workflows
