+++ invenio/modules/workflows/widgets/bibmatch_widget.py
@@ -17,7 +17,7 @@
 ## along with Invenio; if not, write to the Free Software Foundation, Inc.,
 ## 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
 
-from invenio.bibworkflow_hp_field_widgets import bootstrap_accept
+from ..hp_field_widgets import bootstrap_accept
 from wtforms import SubmitField, Form
 __all__ = ['bibmatch_widget']
