+++ modules/Makefile.am
@@ -38,7 +38,6 @@ SUBDIRS = bibauthorid \
           bibsort \
           bibsword \
           bibupload \
-          bibworkflow \
           elmsubmit \
           miscutil \
           oaiharvest \
