+++ modules/bibworkflow/lib/bibworkflow_regression_tests.py
@@ -20,11 +20,11 @@
 BibWorkflow Unit tests - functions to test workflows
 """
 
-from invenio.inveniomanage import db
+from invenio.ext.sqlalchemy import db
 from invenio.testsuite import (make_test_suite,
                                run_test_suite,
                                FlaskSQLAlchemyTest)
-from invenio.bibworkflow_config import CFG_OBJECT_VERSION
+from invenio.modules.workflows.config import CFG_OBJECT_VERSION
 
 
 class TestWorkflowStart(FlaskSQLAlchemyTest):
