+++ modules/bibworkflow/lib/workflows/Makefile.am
@@ -15,6 +15,6 @@
 ## along with Invenio; if not, write to the Free Software Foundation, Inc.,
 ## 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
 
-pylibdir = $(libdir)/python/invenio/bibworkflow_workflows
+pylibdir =
 
 pylib_DATA =
