+++ modules/webdeposit/lib/webdeposit_regression_tests.py
@@ -252,7 +252,7 @@ class TestWebDepositUtils(InvenioTestCase):
         from invenio.cache import cache
         from invenio.config import CFG_PREFIX
         from invenio.modules.workflows.models import Workflow
-        from invenio.bibworkflow_config import CFG_WORKFLOW_STATUS
+        from invenio.modules.workflows.config import CFG_WORKFLOW_STATUS
         from invenio.legacy.bibsched.scripts.bibsched_model import SchTASK
 
         from invenio.webdeposit_utils import get_form, create_workflow, \
